Send each email inside a transaction instead of all them in one transaction
We used to have one transaction for all our emails. This is incorrect -- if we
send 50 emails, and fail on the 51st, all 51 would not be marked as sent and
tried again.

This way we mark all successful emails as sent, but still fail to mark ones that
throw an exeption