Fixes Advisory overprocessing bug re: #171
When we switched criticalities to enum, we inadvertently introduced an
issue where the advisory importer would flag every single advisory has
requiring importing *every single time*.

Issue had to do with how enum values are retrieved from the database;
AdvisoryImporter#has_changed? would always fail. This fixes that,
and adds a test so we can detect it in the future.