Speedup the package merge function
We were spending too much time loading packages for each bundle when making the
bundlepackages a set, and this timed out the transaction.

Now we just check to see if the bundle has the "canonical package" and add it if
it doesn't.

What this doesn't do is solve the problem of duplicate bundledpackages as
opposed to duplicate *packages*. That is, if we have more than one
bundledpackage for the same bundle_id,package_id, we won't delete it here. This
is a separate problem anyway, and it was bad to merge them together. As
discussed we need to add an uniq constraint and some kind of functionality to
make sure we don't do this.

This PR is now limiting the merge_dupe_packages to just merge duplicate packages
that share a name,version,release,arch.

Closes #270