Refactor version constraint checks
This removes some specificity from the programming interface, but I actually
think it adds clarity in the process. Though these three operations were
identical, they were named differently I think for (a) historical reasons
and (b) for contextual clarification. But I think this actually adds to the
cognitive load. Now it's clear exactly what they do, and reading the calling
code (the only place they were called was `affects?` in the VD model) is now
easier to reason about because it doesn't require drilling into the
implementation of the called package methods.

I'm totally willing to be convinced otherwise though.