Improve documentation and local setup
The setup process has been improved to make it easier to get the project
up and running. The default install script in `bin/setup` has been
extended to cover the required setup steps, and the README has been
rewritten to include up-to-date information on the project and its
requirements.

Furthermore, the seeds have been fixed and include a working user
account now. This makes it possible to run the setup script and
immediately log into the application.